Extract QF reform in independent file
